segundo commit

+++ src/main/java/acc/br/desafiofullstack/services/FornecedorPessoaFisicaService.java
@@ -1,37 +1,66 @@
 package acc.br.desafiofullstack.services;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
 import org.hibernate.HibernateException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
+import org.springframework.stereotype.Service;
 
+import acc.br.desafiofullstack.model.Fornecedor;
 import acc.br.desafiofullstack.model.FornecedorPessoaFisica;
 import acc.br.desafiofullstack.repository.FornecedorPessoaFisicaRepository;
-
+import acc.br.desafiofullstack.repository.FornecedorRepository;
+import acc.br.desafiofullstack.utils.ValidadorCampos;
+@Service
 public class FornecedorPessoaFisicaService {
        
     @Autowired
     FornecedorPessoaFisicaRepository fornecedorPFRepository;
+    @Autowired
+    FornecedorRepository fornecedorRository;
 
-    public FornecedorPessoaFisica createFornecedorPF(FornecedorPessoaFisica fornecedorPF, String rg) throws Exception {
+    ValidadorCampos validadorCampos = new ValidadorCampos();
+    
+    public Optional<Fornecedor> createFornecedorPF(FornecedorPessoaFisica fornecedorPF, long id) throws Exception {
         try {
-            Optional<FornecedorPessoaFisica> fornecedorPFExist = fornecedorPFRepository.findByRg(rg);
 
-            if (fornecedorPFExist.isPresent()) {
-                throw new Exception("Fornecedor PF CPF já existente!");
+            String rg = fornecedorPF.getRg();
+            LocalDate dataNascimento = fornecedorPF.getDataNascimento();
+           
+            validadorCampos.validateRG(rg);
+            validadorCampos.validateDataNascimento(dataNascimento);
+
+            Optional<FornecedorPessoaFisica> fornecedorPFRgExist = fornecedorPFRepository.findByRg(rg);
+            Optional<FornecedorPessoaFisica> fornecedorPFExistWithFornecedor = fornecedorPFRepository.findByFornecedorId(id);
+
+            if (fornecedorPFRgExist.isPresent()) {
+                throw new Exception("Fornecedor PF RG já vinculado a um fornecedor");
             }
 
-            fornecedorPFRepository.save(fornecedorPF);
+            if (fornecedorPFExistWithFornecedor.isPresent()) {
+                throw new Exception("Fornecedor já possui RG vinculado!");
+            } 
 
-            return fornecedorPF;
+            Optional<Fornecedor> fornecedorExist = fornecedorRository.findById(id);
+
+            if (!fornecedorExist.isPresent()) {
+                throw new Exception("Fornecedor não encontrado!");
+            }
+
+            fornecedorPF.setFornecedor(fornecedorExist.get());
+            fornecedorExist.get().setPessoaFisica(fornecedorPF);
+            fornecedorRository.save(fornecedorExist.get());
+
+            return fornecedorExist;
         } catch (HibernateException e) {
             throw new Exception("Erro na execução do Hibernate: " + e.getMessage());
         } catch (DataAccessException e) {
             throw new Exception("Erro de acesso aos dados: " + e.getMessage());
         } catch (Exception e) {
-            throw new Exception("Ocorreu um erro inesperado: " + e.getMessage());
+            throw new Exception(e.getMessage());
         }
     }
 
@@ -50,7 +79,7 @@ public class FornecedorPessoaFisicaService {
         } catch (DataAccessException e) {
             throw new Exception("Erro de acesso aos dados: " + e.getMessage());
         } catch (Exception e) {
-            throw new Exception("Ocorreu um erro inesperado: " + e.getMessage());
+            throw new Exception(e.getMessage());
         }
     }
 
@@ -76,28 +105,33 @@ public class FornecedorPessoaFisicaService {
         } catch (DataAccessException e) {
             throw new Exception("Erro de acesso aos dados: " + e.getMessage());
         } catch (Exception e) {
-            throw new Exception("Ocorreu um erro inesperado: " + e.getMessage());
+            throw new Exception(e.getMessage());
         }
     }
 
-    public String getFornecedorPF(long id) throws Exception {
+    public Optional<FornecedorPessoaFisica> getFornecedorPF(long id) throws Exception {
         try {
             Optional<FornecedorPessoaFisica> fornecedorPF = fornecedorPFRepository.findById(id);
             if (!fornecedorPF.isPresent()) {
                 throw new Exception("Fornecedor PF não encontrado!");
             }
-            return fornecedorPF.get().toString();
+            return fornecedorPF;
         } catch (HibernateException e) {
             throw new Exception("Erro na execução do Hibernate: " + e.getMessage());
         } catch (DataAccessException e) {
             throw new Exception("Erro de acesso aos dados: " + e.getMessage());
         } catch (Exception e) {
-            throw new Exception("Ocorreu um erro inesperado: " + e.getMessage());
+            throw new Exception(e.getMessage());
         }
     }
 
     public String deleteFornecedorPF(long id) throws Exception {
         try {
+            Optional<FornecedorPessoaFisica> fornecedorPFExist = fornecedorPFRepository.findById(id);
+
+            if (!fornecedorPFExist.isPresent()) {
+                throw new Exception("Fornecedor PF não encontrado!");
+            }
             fornecedorPFRepository.deleteById(id);
 
             return "Fornecedor PF" + id + "deletado.";
@@ -106,7 +140,7 @@ public class FornecedorPessoaFisicaService {
         } catch (DataAccessException e) {
             throw new Exception("Erro de acesso aos dados: " + e.getMessage());
         } catch (Exception e) {
-            throw new Exception("Ocorreu um erro inesperado: " + e.getMessage());
+            throw new Exception(e.getMessage());
         }
     }
 }
